+++ src/rules/rules.ts
@@ -14,6 +14,8 @@ import { compareRequestRequiredRule } from "./compare-request-required.js";
 import { ignoreSwaggerDefinitionsRule } from "./ignore-swagger-definitions.js";
 import { compareXMsIdentifiersRule } from "./compare-x-ms-identifiers.js";
 import { ignoreXMsExamplesRule } from "./ignore-x-ms-examples.js";
+import { pathAddedRule } from "./path-added.js";
+import { pathRemovedRule } from "./path-removed.js";
 
 /** Determines whether a diff rule applies and confirms an allowed or disallowed scenario. */
 export enum RuleResult {
@@ -47,6 +49,8 @@ export function getApplicableRules(args: any): RuleSignature[] {
     flagTypeDifferencesRule,
     compareResponseRequiredRule,
     compareRequestRequiredRule,
+    pathAddedRule,
+    pathRemovedRule,
   ];
 
   const preserveDefinitions = args["preserve-definitions"];